Change how player goes to falling state
Player now keeps track of if he pressed the jump button
in order to determine wheter he is in the air or not.

This was done because when player's stood on a slope, the
previous method of using the player's velocity to check if he
was falling would always register him as falling when he started
moving. This was because when moving up or down a slope, his y velocity
would not be zero, and this would cause him to play the jump animation.