changed spritedisplaybox sprite
Remove the heart, ammo, and money sprites from the
sprite display box sprite and make them into their
own sprites. This way you can draw the individual
sprites on top of the display box, and when
colouring the display box with the team color, it doesn't
color the sprites and make them look odd.